+++ ntex-net/src/uring/io.rs
@@ -30,7 +30,7 @@ struct HandleWrapper(WeakStreamCtl);
 impl Handle for HandleWrapper {
     fn query(&self, id: any::TypeId) -> Option<Box<dyn any::Any>> {
         if id == any::TypeId::of::<types::PeerAddr>() {
-            let addr = self.0.with_io(|io| io.peer_addr().ok());
+            let addr = self.0.with_io(|io| io.peer_addr().ok()).flatten();
             if let Some(addr) = addr.and_then(|addr| addr.as_socket()) {
                 return Some(Box::new(types::PeerAddr(addr)));
             }